% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_group.R
\name{get_age_group}
\alias{get_age_group}
\title{Validate and Convert Age Group to Census Codes}
\usage{
get_age_group(age_group = "total")
}
\arguments{
\item{age_group}{Character or numeric. Acceptable values include:
\itemize{
\item Keywords from the dictionary above.
\item Open-ended notation like \code{"65plus"}.
\item Ranges like \code{"15to69"} (validated against Census bins).
\item Numeric codes 0–18 (see dictionary).
}}
}
\value{
A numeric vector of valid census AGEGROUP codes.
}
\description{
A helper function that checks whether an age group keyword or numeric
code is valid according to Census definitions, and converts it to the
corresponding numeric AGEGROUP code(s).
}
\details{
Dictionary of valid age groups (Source: U.S. Census Bureau, Population Division,
Release date: June 2024):

\itemize{
\item \code{"total"} = 0
\item \code{"0to4"} = 1
\item \code{"5to9"} = 2
\item \code{"10to14"} = 3
\item \code{"15to19"} = 4
\item \code{"20to24"} = 5
\item \code{"25to29"} = 6
\item \code{"30to34"} = 7
\item \code{"35to39"} = 8
\item \code{"40to44"} = 9
\item \code{"45to49"} = 10
\item \code{"50to54"} = 11
\item \code{"55to59"} = 12
\item \code{"60to64"} = 13
\item \code{"65to69"} = 14
\item \code{"70to74"} = 15
\item \code{"75to79"} = 16
\item \code{"80to84"} = 17
\item \code{"85plus"} = 18
}
}
\examples{
library(gdp)
get_age_group("65plus")
get_age_group("65to79")
get_age_group(1)

}
