% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{get_rates}
\alias{get_rates}
\title{Prepare Event Rates with Boundaries}
\usage{
get_rates(
  rate,
  population,
  rate_per = 1000,
  state = NULL,
  county = NULL,
  time = 1
)
}
\arguments{
\item{rate}{List with event rate \code{event_rate} and \code{GEOID}.}

\item{population}{List with \code{GEOID} and estimated population \code{count}
for the target rate.}

\item{rate_per}{Rate of events per unit population (e.g., per 1,000).
Defaults to \code{1000}.}

\item{state}{The state for which the data is requested. State names, postal codes,
and FIPS codes are accepted. Defaults to \code{NULL}.}

\item{county}{The county for which the data is requested. FIPS
codes are accepted. Defaults to \code{NULL}.}

\item{time}{Number of time periods to generate. Defaults to \code{1}.}
}
\value{
A list with:
\itemize{
\item \code{data} A data frame of counties with event rates, population,
and estimated counts.
\item \code{boundary} An \code{sf} object of the requested state or county
boundary.
}
}
\description{
Combines event rates, population, and state or county boundary into a single object for
simulation.
}
\examples{
data(Stroke_Rate)
pop<-get_census_population(geography="county",age_group = "65plus")
get_rates(rate = Stroke_Rate,population=pop)
get_rates(rate = Stroke_Rate,population=pop,state = "01",county = "001")

}
