% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Weighted_Stroke_Events_Function.R
\name{GetStrokes}
\alias{GetStrokes}
\title{A single simulated set of Stroke Events for specific state_code or county_code.}
\usage{
GetStrokes(
  stroke_rate,
  population,
  density,
  counties,
  state_code = FALSE,
  county_code = FALSE,
  warn = TRUE,
  years = 1,
  ...
)
}
\arguments{
\item{stroke_rate}{Data frame with Hospitalization Average Rate per 1,000 Medicare Beneficiaries and cnty_fips (2014-2020) (Data from CDC).}

\item{population}{Data frame with population by cnty_fips ages 65+ (Census Data).}

\item{density}{RasterLayer with population density in CONTINENTAL U.S. excluding AK.}

\item{counties}{shape file, geometry of counties (Census data, ONLY CONTINENTAL U.S. excluding AK).}

\item{state_code}{State Code, example: "01".}

\item{county_code}{county_code, example: "003".}

\item{...}{Further arguments passed to exact_extract in the default setup.}
}
\value{
A data frame including

\code{longitude} Coordinates,

\code{latitude} Coordinates,

\code{stroke_type} ,

\code{category} ,

\code{cnty_fips} County fip,

\code{display_name} County name,
}
\description{
Random Strokes
}
\details{
Random Strokes
}
\author{
Ferney Henao-Ceballos
}
