% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroke_hospitalization}
\alias{stroke_hospitalization}
\title{Stroke Hospitalization Rates}
\format{
A data frame with 3226 rows and 3 columns:
\describe{
\item{display_name}{county name.}
\item{event_rate}{Stroke rate.}
\item{GEOID}{Unique geographic identifier.}
}
}
\usage{
stroke_hospitalization
}
\description{
Data from the U.S. Centers for Disease Control and Prevention (CDC) Interactive
Atlas of Heart Disease and Stroke with annual stroke hospitalization rates for
medicare beneficiaries aged 65 and older at the county level from 2014 to 2020.
}
\keyword{datasets}
