% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroke_mortality}
\alias{stroke_mortality}
\title{Stroke Mortality Rates}
\format{
A data frame with 51 rows and 4 columns:
\describe{
\item{event_rate}{Stroke rate.}
\item{GEOID}{Unique geographic identifier.}
\item{NAME}{State name.}
}
}
\usage{
stroke_mortality
}
\description{
Data from the U.S. Centers for Disease Control and Prevention (CDC) with mortality rate per 100,000 total population at the state level for 2023.
}
\keyword{datasets}
