% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stroke_Rate}
\alias{Stroke_Rate}
\title{Example dataset: Stroke_Rate}
\format{
A data frame with stroke rates by county
}
\usage{
Stroke_Rate
}
\description{
Example dataset: Stroke_Rate
}
\keyword{datasets}
