% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_event_locations.R
\name{generate_event_locations}
\alias{generate_event_locations}
\title{A Simulated Set of Events for a Specific State or County.}
\usage{
generate_event_locations(
  rate,
  rate_per,
  population,
  family = c("poisson", "negative_binomial"),
  control = list(),
  state = NULL,
  county = NULL,
  labels = NULL,
  probs = NULL,
  time = 1,
  warn = TRUE,
  ...
)
}
\arguments{
\item{rate}{Data frame with event rate \code{event_rate} and county FIP \code{county_fip}.}

\item{rate_per}{Numeric. Rate of events per unit population (e.g., per 1.000).}

\item{population}{List with \code{GEOID} and estimated population \code{count}
for the target rate.}

\item{family}{A distribution family used to generate event counts. Options include \code{poisson} or \code{negative_binomial}.}

\item{control}{A list of parameters controlling the chosen distribution (e.g., dispersion for \code{negative_binomial}).}

\item{state}{The state for which the data is requested. State names, postal codes,
and FIPS codes are accepted. Defaults to \code{NULL}.}

\item{county}{The county for which data is requested. FIPS
codes are accepted. Defaults to \code{NULL}.}

\item{labels}{Label assigned to each generated event.}

\item{probs}{Numeric vector of probabilities corresponding to \code{labels}. If omitted, labels are sampled uniformly.}

\item{time}{Number of time periods to generate. By default 1.}

\item{warn}{Logical. If TRUE, display warnings during simulation.}

\item{...}{Additional arguments passed to \code{tigris::counties()} and \code{geodata::population()}.}
}
\value{
A data frame of generated locations with:
\itemize{
\item \code{longitude} Longitude.
\item \code{latitude} Latitude.
\item \code{county_fip} Unique county-level geographic identifier.
\item \code{label} Label assigned to each generated event.
}
}
\description{
A Simulated Set of Events for a Specific State or County.
}
\details{
We generate realistic in-silico cohorts of events using rates at the county level.
Counts for each county are computed as: \code{time * event_rate *  count / rate_per}

Counts can be generated using different statistical distributions via the \code{family} argument,
including \code{poisson} or \code{negative binomial}, with control over
distribution-specific parameters (e.g., dispersion for negative binomial distribution \code{size}).

The events are distributed within counties in proportion to estimated population counts.
Latitude and longitude of each event are generated randomly, weighted by the GPWv4 population density map,
so denser areas are more likely to receive events, while sparsely populated areas receive fewer. The algorithm
also accounts for pixel area to avoid bias due to differing cell sizes.

A recursive accept-reject algorithm based on county or state boundaries ensures that all
generated points fall inside the target polygons while preserving the target counts and proportions.
}
\examples{
library(gdp)
data(Stroke_Rate)
population<-get_census_population(geography="county")
strokes<-generate_event_locations(rate=Stroke_Rate,
                                  rate_per=1000,
                                  population = population,
                                  family = "negative_binomial",
                                  control = list(size=10),
                                  state = "01")


}
\references{
\insertRef{Rpack:bibtex}{Rdpack}

\insertRef{R-exactextractr}{gdp}

\insertRef{R-purrr}{gdp}

\insertRef{R-tigris}{gdp}

\insertRef{R-geodata}{gdp}

\insertRef{R-dplyr}{gdp}
}
