% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_population.R
\name{get_census_population}
\alias{get_census_population}
\title{Population Estimates by Age Group from Census Bureau}
\usage{
get_census_population(
  geography,
  breakdown = "AGEGROUP",
  age_group = "65to84",
  state = NULL,
  county = NULL,
  geometry = TRUE,
  ...
)
}
\arguments{
\item{geography}{The geography of your data. Available geographies for the most
recent data vintage are listed \href{https://api.census.gov/data/2019/pep/population/geography.html}{here}. "cbsa" may be used an alias for "metropolitan
statistical area/micropolitan statistical area".}

\item{breakdown}{The population breakdown(s).
Acceptable values are \code{AGEGROUP}, \code{"RACE"},
\code{"SEX"}, and \code{"HISP"}.}

\item{age_group}{Character keywords (e.g., \code{"65plus"},
\code{"15to69"}) or numeric code as defined by the U.S. Census Bureau. See arguments for
\code{get_age_group} function.}

\item{state}{The state for which the data is requested. State names, postal codes, and FIPS
codes are accepted. Defaults to NULL.}

\item{county}{The county for which the data is requested. County names and FIPS codes
are accepted. Must be combined with a value supplied to ‘state‘. Defaults to NULL.}

\item{geometry}{if FALSE (the default), return a regular tibble of ACS data. if TRUE, uses
the tigris package to return an sf tibble with simple feature geometry in the
‘geometry‘ column.}

\item{...}{Additional arguments passed to \code{tidycensus::get_estimates()}.}
}
\value{
A list of data frames, one per strata. Each data frame contains:
\describe{
\item{GEOID}{Unique geographic identifier.}
\item{Name}{County or state name.}
\item{population}{Estimated population for the given strata.}
\item{age_group}{Age group.}
}
}
\description{
A helper function that uses \code{tidycensus::get_estimates()} to obtain
U.S. Census Bureau Population Estimates data.The functions allows filtering by
age group, race, sex, and Hispanic origin.
}
\examples{
get_census_population(geography="county",breakdown=c("AGEGROUP"))
get_census_population(geography="county",breakdown=c("AGEGROUP","SEX"))
get_census_population(geography="county",breakdown=c("AGEGROUP","SEX"),state="01", county ="001")
}
\references{
\insertRef{Rpack:bibtex}{Rdpack}

\insertRef{R-dplyr}{gdp}

\insertRef{R-tidycensus}{gdp}
}
