% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_population.R
\name{get_census_population}
\alias{get_census_population}
\title{Population Estimates by Age Group from Census Bureau}
\usage{
get_census_population(
  geography = c("state", "county", "cbsa", "combined statistical area"),
  breakdown = "AGEGROUP",
  age_group = "Total",
  state = NULL,
  county = NULL,
  geometry = TRUE,
  vintage = 2024,
  progress = FALSE,
  ...
)
}
\arguments{
\item{geography}{The geography of your data.

Available geographies include: 'state', 'county', 'cbsa', 'combined statistical area'

"cbsa" is used as alias for "metropolitan statistical area/micropolitan statistical area".}

\item{breakdown}{The population breakdown(s). Acceptable values are \code{"AGEGROUP"}, \code{"RACE"},
\code{"SEX"}, and \code{"HISP"}.}

\item{age_group}{Character keywords (e.g., \code{"65plus"},
\code{"15to69"}) or numeric code as defined by the U.S. Census Bureau. See arguments for
\code{get_age_group} function.}

\item{state}{The state for which the data is requested. State names, postal codes, and FIPS
codes are accepted. Default to NULL.}

\item{county}{The county for which the data is requested. County FIP codes
are accepted.}

\item{geometry}{If FALSE (the default), return a regular tibble. if TRUE, uses
the tigris package to return an sf tibble with simple feature geometry in the
‘geometry‘ column.}

\item{vintage}{Vintage year; see tidycensus for details.}

\item{progress}{By default FALSE.}

\item{...}{Additional arguments passed to \code{tidycensus::get_estimates()}.}
}
\value{
A list of data frames, one per strata. Each data frame contains:
\itemize{
\item \code{geoid:} Unique geographic identifier.
\item \code{name:} County or state name.
\item \code{population:} Estimated population for the given strata.
\item \code{age_group:} Age group.
}
}
\description{
A helper function that uses \code{tidycensus::get_estimates()} to obtain
U.S. Census Bureau Population Estimates. The function allows filtering by
age group, race, sex, and Hispanic origin.
}
\examples{
get_census_population(geography="county")
get_census_population(geography="county",breakdown=c("AGEGROUP","SEX"))
get_census_population(geography="state",breakdown=c("AGEGROUP","SEX"),state="Alabama")
get_census_population(geography="county",breakdown=c("AGEGROUP","SEX"),state="01", county ="001")
get_census_population(geography="cbsa",breakdown=c("AGEGROUP","SEX"))
get_census_population(geography="combined statistical area")

}
\references{
\insertRef{Rpack:bibtex}{Rdpack}

\insertRef{R-dplyr}{pointgen}

\insertRef{R-tidycensus}{pointgen}
}
