% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{Get Population Density Values}
\usage{
get_density(
  geography = c("state", "county", "zcta", "cbsa", "combined statistical area"),
  state = NULL,
  county = NULL,
  year = 2020,
  res = 0.5,
  path = tempdir(),
  ...
)
}
\arguments{
\item{geography}{Character string specifying the boundary type from the tigirs package.

Available geographies include: 'state', 'county', 'cbsa', 'zcta', 'combined statistical area'

"cbsa" is used as alias for "metropolitan statistical area/micropolitan statistical area".}

\item{state}{The state for which the data is requested. State names, postal codes,
or FIPS codes are accepted. Default to \code{NULL}.}

\item{county}{The county for which the data is requested. FIPS codes
are accepted. Default to \code{NULL}.}

\item{year}{Year of the population density data. One of 2000, 2005, 2010, 2015, 2020. Default is \code{2020}.}

\item{res}{Resolution of the raster, validvalues are 10, 5, 2.5, and 0.5 (minutes of a degree). Default is \code{0.5}.}

\item{path}{Character. Directory path to save the downloaded raster.
If \code{NULL}, a temporary directory is used.}

\item{...}{Additional arguments passed to \code{geodata::population()}.}
}
\value{
Extracted values of cells from a raster (e.g., population density) within the
boundary of a specified state or county.
}
\description{
Get Population Density Values
}
\details{
This function retrieves global population density rasters from
the \pkg{geodata} package. Data is obtained from the GPWv4 dataset.
}
\examples{
# Get 2020 population density at 0.5 resolution
get_density(geography="state")
get_density(geography="state",state="01")
get_density(geography="county",state="01",county="001")

}
